fix(blog): restore /categories index route
Header nav points at /categories while the index lived at /category.
Serve the index at /categories, keep /category/$category slugs, and
redirect exact /category for old links.

Closes #1469
