+++ apps/blog/public/sitemap.xml
@@ -1341,7 +1341,7 @@
     <lastmod>2026-08-22</lastmod>
   </url>
   <url>
-    <loc>https://blog.duyet.net/category</loc>
+    <loc>https://blog.duyet.net/categories</loc>
     <lastmod>2026-08-22</lastmod>
   </url>
   <url>
