+++ apps/blog/scripts/generate-static-files.ts
@@ -390,7 +390,7 @@ ${categories
     <lastmod>${new Date().toISOString().split("T")[0]}</lastmod>
   </url>
   <url>
-    <loc>${SITE_URL}/category</loc>
+    <loc>${SITE_URL}/categories</loc>
     <lastmod>${new Date().toISOString().split("T")[0]}</lastmod>
   </url>
   <url>
